Revert "Kernel/PCI: Allow to set the PCI IRQ line of a device"
This reverts commit 36a82188a88c95315e03f6fcede237bc66831702.

This register is write-only for the firmware (BIOS), and read-only for
us so we shouldn't set the PCI IRQ line never.
The firmware figured out the IRQ routing to the PIC for us, so changing
it won't affect anything. I was mistaken when I thought that changing
the value of this register will allow us to change its interrupt line,
like when changing a PCI BAR to relocate device resources as desired
with the requirements of the OS.