Utilities: Clean up CMakeLists.txt
We previously had a lot of duplication (and mismatched settings) around
whether a target was "special" or not, although the only special thing
about it was that its name was a reserved keyword and therefore had to
be renamed later.

Fix that by mostly consolidating the different configuration paths and
then purging all duplicated linking libraries from the bottom of the
file.