LibPDF: Add some scaffolding for the `sh` operator
Adds a `Shading` class, which is supposed to be implemented similarly
to LibPDF's ColorSpace class: `Shading::create()` will return a Shading
subclass for every shading type, and they'll all implement some future
pure virtual method for drawing the shading.

Shadings are used with both the `sh` operator and as part of pattern
color spaces. We don't implement the latter yet either, and `Shading`
will assume that its used with the `sh` operator for now.

For now, `Shading::create()` returns a `rendering_unsupported_error`
for every shading type, so this commit doesn't do a whole lot.
It does produce different errors for different shading types though,
making it possible to see which shading types are common.

It also removes the last use of `RENDERER_TODO`, so remove that :^)