LibWeb: Empty CE reaction queue instead of destroying it on exception
If an exception occurs in a custom element constructor, we clear the
reaction queue by destroying it, instead of emptying the Vector.
https://github.com/SerenityOS/serenity/blob/3da6916383afc1c8a8b59dae38c4bc02680b50ba/Userland/Libraries/LibWeb/DOM/Element.cpp#L2033

This causes a UAF here, as async upgrades (i.e. custom elements not
created by document.createElement) are performed in this loop:
https://github.com/SerenityOS/serenity/blob/3da6916383afc1c8a8b59dae38c4bc02680b50ba/Userland/Libraries/LibWeb/Bindings/MainThreadVM.cpp#L657

Fixes crash when loading https://github.com/SerenityOS/serenity