CatDog: Fix sleeping animation not playing
Previously, you'd see one frame of the sleeping animation, and then it
would stop. This was because Frame2 = 0x1 and Frame1 = 0x0, so once the
Frame2 bit was set, or-ing with Frame1 would not unset it. Since the
current frame was already tracked separately to the state, we can avoid
setting it in the state, and instead only use it when we fetch the
bitmap, preventing this issue.