LibGfx/WebPWriterLossless: Write smaller grayscale images
We used to write a color indexing transform for grayscale images.
This stores a palette and then indexes into that palette.

For grayscale images with constant alpha, we don't need to store
the palette image: Using a subtract green transform has the same effect.

(For animations, most frames don't have constant alpha because
AnimationWriter replaces identical pixels with transparent black,
making sure that these frames have a mix of opaque and transparent
pixels. But the first frame of a grayscale animation will use this.)

Only saves a couple of bytes for storing the palette image,
but it's also free in term of performance, and it's conceptually
pleasing.

Reduces the size of 7z7c.webp by 30 bytes, from 8818 to 8788 bytes.