LibGfx/JPEGXL: Add support for decoding embedded ICC profiles
The test case was generated by opening `modular_property_8.jxl` in
Krita, then changing the Color Profile to Apple RGB and then exporting
as PNG. Finally, the conversion to JPEGXL was made with:
`cjxl --container=0 --modular=1 -d 0 -e 9 file.png icc.jxl`

Note that, we can't use Krita to export a jxl as it always create a
container instead of a raw stream. Also, this needs an old version of
`cjxl` (I used 0.7), as more recent version reencode the ICC profile
using the JPEGXL's internal representation for color profile, which is
encoded differently.