LibWeb: Begin implementing FontFaceSet.prototype.load
This implementation is incomplete in that we do not fully implement the
steps to match the given font against the fonts in the set.

This is used by fonts.google.com to load the fonts used for sample text.

(cherry picked from commit 9bdf2e928c448585a6349bef36d7cb98ccc0607b)