WindowServer: Handle shifted submenus during safe submenu navigation
The old calculation was unnecessarily complicated and didn't handle
horizontally or vertically shifted menus correctly.

To determine the two triangle corners at the submenu, we can simply
use the submenu's rect() and subtract the position of the parent menu
to get the correct relative coordinates.
This fixes the case when the menu is shifted vertically.

If the submenu is shifted horizontally, we need to use the opposite
corners of the window rect.