LibC: Don't push a fake return address in `create_thread` on non-x86
The AArch64 and RISC-V ABI use a caller-saved register to store the
return address instead of (always) storing it on the stack.
All non-special registers of newly created threads are zero by default,
so we don't need to anything special for those architectures here.

This change also causes the stack pointer to no longer be misaligned for
secondary threads on those architectures.