LibGfx/JPEG2000: Reject images with bit_depth != 8
We should upsample (if < 8) or downsample (if > 8), but we were just
passing the data unmodified. That looked wrong; now we instead emit
an error.

(In practice, this seems to happen extremely rarely. I'm aware of a
single PDF that contains a 4 bits-per-component image, and that's
palettized and the PDF wants the 4 bits expanded to 8 bits. So
we were accidentally doing the right thing there actually. But let's
intentionally do the right thing later. And we were doing the wrong
thing when that file was viewed outside a PDF.)