Kernel/Memory: Ignore reserved range at 0x000003ff f000 0000
On Meteor Lake this range covers the GFX address space reservations,
at a very high address. Nothing actually ends up using it, so we can
ignore it. Moreover if we don't we end up with highest_address at the
4 TiB mark which would need an enormous page table to map this
huge (0-4TiB) range. Until we support mapping I/O space piecewise we
need to skip this range.