Meta: Smarter "apparent size" flag detection for `du`
The previous code assumed that, under a Darwin system, you would either
have the GNU coreutils version of `du` installed through brew, which
adds a 'g' prefix, or the default, BSD-like version.

This assumption can be easily broken by simply directly installing GNU's
coreutils or through a package manager that doesn't prepend a 'g' to it.

We now instead always use whatever `du` is in the user's $PATH and try
to detect which flavor it is and, consequently, which flag should be
used to retrieve the "apparent size" of something in disk.

The heuristic is not perfect and may lead to the wrong flag being
selected in edge cases, but it can always be amended in the future.