LibGfx/JPEG2000: Support standalone palettized jpeg2000 files
The approach is to make a distinction between components and channels
(mirroring spec language). If there's no palette, each component
directly maps to a channel. (This uses move semantics, so there's
next to no data copying happening.) If there's a palette, it uses
the component mapping box to map components to channels, likely using
the palette for some channels. So a single component of palette indices
is converted to N (with N usually 3, sometimes 4) full channels.

If a palette is present, this copying has some overhead,
but it's conceptually simple. Things that should just work:

* Palette indices of any bitness

* Palette values with bpp in [1..16] (existing support handles)

* Two-component images where one component is a palette index and
  the other is an alpha channel (don't know how to make those, though)

* Palettes for CMYK files (and for other color spaces)

Note that this codepath is only used for standalone palettized JPEG2000
flies, not for JPEG2000 files in PDFs.

(To create ref-indexed.png, I opened
openjpeg-lossless-indexed-u8-rgb-u8.jp2 in Photoshop and saved it as
PNG.)