LibPDF: Skip whitespace before values in compressed stream
`parse_value()` expects to start on the value.

It's a bit surprising that the object_offset in the compressed stream
would point at whitespace instead of the value directly, but it seems to
happen in at least one file: This allows us to open 0001215.pdf from the
CC-MAIN-2021-31-PDF-UNTRUNCATED test corpus.