LibGfx/JBIG2: Handle negative offsets in text region decoding better
Table 12 in the spec says to use `⌊RDW_I / 2⌋ + RDX_I`. We were just
using `RDW_I / 2 + RDX_I`, but C's division semantics are different
from flooring for negative numbers. Use floor_div() instead.

(I looked through other uses of ` / ` in this file; this was the only
instance of this problem I found.)

In addition to fixing the test in our repo, also fixes an "infinite"
loop in 0001564.pdf page 20, where due to this mistake we'd get the
wrong delta width / heights for a later bitmap, which led to a
`width + delta_width` computation producing -1 but being stored in a
`u32`, and decoding UINT_MAX rows takes a while.  (Our arithmetic
decoder never runs out of data.)