Tests/LibPDF: Tweak Gouraud shading tests to use nonlinear functions
The spec says that one should linearly interpolate the function
input and then do a per-pixel function evaluation (by sampling the
function into a 1D texture and do a per-pixel lookup I suppose).

If the function is linear, just evaluating the function at the
vertices and interpolating the output is equivalent, so use a
nonlinear function to test for this.