LibPDF: Implement support for clipping to arbitrary paths
While a clip is active, we now paint into a single additional bitmap
(m_clipped_bitmap, allocated once at first use, same size as the output
bitmap, but has a clip rect equal to the clip's bounding box applied for
efficiency). When the clip is popped, or when a new clip becomes active,
we paint this background buffer to the main output, using the clip path
as alpha channel.

The clip path is rasterized eagerly to graphics state.

m_clipped_bitmap_painter and m_clipped_bitmap_anti_aliasing_painter
are non-empty if a clip is active, and painter()  and
anti_aliasing_painter() return those in that case.

(m_clipped_bitmap can be non-null even if there's no clip, since
it's allocated when the first clipping path is added and then kept
around for future clips.)

The rasterized clipping path is state().clipping_state.clip_path_alpha.

Transparent contents work by initializing the clip color buffer to the
current page contents each time a new clip is added.

Not super duper optimized, but also not super expensive: ~5-10%
overhead compared to before (for pages with complex clips), and
a big progression visually.