LibPDF+Clients: Have a single "Apply Clip" debug option
This replaces the granular "Clip Images", "Clip Paths", and
"Clip Text" debug options with a single "Apply Clip" option
that controls all clipping.

The main motivation is that support for clipping to paths (instead of
the current implementation that only clips to the clipping paths's
bounding boxes) makes it a bit tricky to have these granular options.
Since this is a debug option, having complexity for this doesn't seem
worth it.

It's also less code.