LibGfx/JBIG2: Implement support for USESKIP in generic regions
As mentioned in #23864, halftone region decoding is the only way
to call the generic region decoding procedure with USESKIP set.

With this, we can decode 25 of the jbig2 files in t89-halftone.zip,
up from 2 before this patch. For example, one file that errored out
previously and decodes fine now is 200-6-45.jb2.