LibGfx/JBIG2: Support striped images with initially unknown height
Do this by just looking through all EndOfStripe segments of the
current page and determining the height that way.

Also add a few consistency checks for PageInformation / EndOfStripe /
EndOfPage / EndOfFile segments. (But there's more to be done;
we could implement 7.3.2.)

With this, we can decode the remaining 15 files (*-stripe.jb2) in
t89-halftone.zip, and we can decode 042_9.jb2 from the power jbig2
test file set.

I haven't seen this in PDFs in the wild.