LibGfx/ImageFormats: Add BilevelImage::create_from_bitmap()
For now, this computes a threshold using Otsu's method, and then
either does global thresholding or Floyd-Steinberg error diffusion
dithering. It converts to grayscale using luminosity, and then
converts that to linear color before doing histogramming, threshold
computation, and dithering.

We should add more dithering methods in the future, and also come
up with a way to make at least some of the dithering methods work
for palettes created by median_cut() instead of just bilevel data,
but for now this is a good start.

This clamps pixel + error to 0..255. I think that shouldn't have
a big visual effect, but it might be worth having two i32 helper
scanlines to check if that looks any different.