Tests/LibGfx: Tweak retained bits in JBIG2 symbol refinement test
Like in the previous commit, this changes the byte at offset 0x30 from
0x00 to 0x01 to set the retention flag of the first symbol segment.

In this file, there's a second symbol segment that refers to the first
one and that exports the first segment's symbols and adds an additional
one to them. Then there's a text segment that refers to just the second
symbol segment.

So this also changes the byte at offset 0x12C from 0x20 to 0x23. That
byte stores the number of referred-to sections in the top 3 bits (this
is the second symbol segment, it refers to one segment, namely the first
symbol segment), the lowest bit stores the retention flag for this
second symbol segment (it needs to be retained since the text segment
will use it), and the second-lowest bit stores if the referred-to
segment should be retained after this segment. Since this segment
re-exports symbols from the first segment, it's still needed, so that
flag is also set.

The text segment then refers to the second symbol segment, and sets the
referred-to segment's retention flag to 0.

This matches how the JBIG2 files in 0000337.pdf (the only file I know of
that uses `SDREFAGG=1`, i.e. symbol segment refining a prior symbol
segment) set their retention bits.

My local change to validate these retention flags a bit does flag that
the two symbol segments were referred to despite not setting their own
retention flag. It does *not* flag that the second symbol segment
cleared the first symbol segment's retention flag despite re-exporting
its symbols.

Interestingly, nothing can set the first symbol segment's retention flag
to 0 after the text segment. The text segment can't do it since it
doesn't directly refer to it. So a decoder that used these retention
flags to keep memory use down would have to keep the first symbol
segment in memory, even though it's no longer used after the text
segment: JBIG2 doesn't seem to have a mechanism to communicate to the
decoder that it's unused.

(In theory, maybe the second symbol segment could not export the first
symbol segment's symbol, then the text segment could explicitly refer to
both, and then it could control the retention flags of both referred-to
segments. But I don't know of any JBIG2 files that set up their symbol
and text segments that way.)