LibGfx/JBIG2: Reject EndOfStripe sequences with decreasing Y coordinate
...instead of the previous behavior, crashing.

This was easy to repro by making a json file with two consecutive
end_of_stripe with decreasing y_coordinate.

This is slightly stricter than the VERIFY() that this replaces in that
it also disallows consecutive stripes with the same Y coordinate.

Also add some spec comments while here.

Having a test for this would require having an invalid JBIG2 file,
something I don't want to create. Maybe we could have `json/invalid`
with just json descriptions of invalid files and check that those are
rejected by the loader. But just saying "the fuzzer will get to this" as
we've done so far seems fine too.