jbig2-from-json: Allow setting referred_to_segments
Segments in a JBIG2 file can refer to other segments. With this,
jbig2-from-json can write these relations between segments.

It's used for refinement regions, text regions, halftone regions,
and custom huffman tables, none of which we can write yet.
This will become useful once we can write one of them, and it's
required for doing so.