LibGfx/JBIG2+jbig2-from-image: Support writing refinement regions
This is similar to the implementation of writing generic regions
in that it duplicates the first half of the refinement region
decoding procedure. We can improve on that in the future.

This doesn't support refining the page itself yet, i.e. a refinement
section without a referred-to segment. Other than that, this is
complete: It supports TPGRON, 7fff stripping, both templates,
etc.