FileManager: Remove unnecessary refresh_tree_view() calls
It's not really needed for most actions: 'Cut' and 'Copy' don't modify
the filesystem in any way, and the 'New File'/'Create Archive' actions
don't even make directories.

The more important thing here is that the TreeViewModel already has a
file watcher that can keep the view in sync for us. :3

So, the only place left that calls it is the 'Show Dotfiles' action,
and that's because the FileSystemModel does not support updates.