Tests/LibGfx: Add a JPEG XL test using the XYB color profile
This color profile is only used for lossy compressions, so I re-encoded
`buggie.png` with a very low distance to simulate a lossless file.

```
$ cjxl -m 1 -d 0.0001 test-inputs/png/buggie.png xyb.jxl
```

The reference decoder outputs at maximum a difference of 1 (per pixel,
the sum of each channel difference) between `xyb.jxl` and `buggie.png`,
so we test against that target for our implementation.