LibGfx/JBIG2+Tests: Add tables C and E, test tables C, E, G, I, L
The new json file is identical to bitmap-symbol-symhuff-texthuff,
except that it sets height and width table selection in the symbol
segment both to 1, and the two s and the one t table in the text
segment all to 1.

PDFium, pdf.js, mutool / jbig2dec can decode this, Preview.app can't.