Tests/LibGfx: Add test for multi-refinement in huffman-coded symbol dict
Similar to refinement in a huffman text region, this decodes fine in
Acrobat Reader, PDFium, and mostly fine in Preview.app (it has the same
artifacts as the other huffman-coded text region files).

pdf.js, and jbig2dec (via mutool) cannot decode it.

I haven't seen this in the wild either, but embedding a huffman
text region stream in a huffman symbol dictionary stream is very
straightforward and there's none of that "arithmetically-coded data
embedded in huffman data" business that we have with refinement in a
huffman region (and that we will have in single-refinement in symbol
dictionaries).

The json file is identical to bitmap-symbol-symhuffrefineseveral.json,
except that `uses_huffman_encoding` is set to true for both symbol and
text regions (not really needed for the first).