LibGfx/JBIG2Writer: Decouple segment data computation from writing
We now store each segment's encoded data on the encoding context.
This will be useful for implementing writing files in the random
access organization, and it will hopefully also be useful in cases
where we need to partially decode the file we're about to write
(for example when refining whole text or halftone regions, or when
refining the current page).

This requires a few more copies than before, but it's small amounts
of data. `compile.sh` time does not change measurably with this.

No behavior change.