Tests/LibGfx: Add JBIG2 test for inner text region composition
Text regions allow setting a background color, and a composition
operator that's used when drawing symbol instances into the text region.
(The final output of the text region is then composited onto the page
with the region's *external* composition operator.)
This adds tests for that.

It follows the same strategy as the compositing tests in #26244:

To test for 'or', paint a region (the mouth) twice in the same location,
on a white background.

To test for 'xor', paint a region (the nose) once in the correct
location, and paint it twice on top of each other at some incorrect
location, on a white background.

To test for 'and', paint an inverted region (the upper eye) twice, on a
black background. To make sure it shows up correctly, draw a black
generic region below the text region, and use an external composition
operator of xor for the two regions, to undo the region being inverted.

To test for 'xnor', paint an inverted region (the lower eye) once in the
correct location, and paint it twice at some other incorrect location,
on a black background. Like for 'and', put a generic black region below.

(Text regions cannot use "replace" for their composition operator.)