imgcmp: Print more details on difference, add --quiet, remove --count
We now print more detail on difference by default:

    % imgcmp serenity-rgb.webp magick-rgb.png
    number of differing pixels: 2261416 (13.48%)
    max error R:    5, G:    4, B:    5
    avg error R: 0.03, G: 0.05, B: 0.07
    max error at (2942, 35): rgb(29, 156, 115) vs rgb(24, 156, 115)
    first difference at (0, 0): rgb(171, 113, 41) vs rgb(170, 113, 42)

Previously with --count, this would print:

    different pixel at (0, 0), rgb(171, 113, 41) vs rgb(170, 113, 42)
    number of differing pixels: 2261416

Locally, I also tried reporting maximum DeltaE, but that takes much
longer to compute. The metrics computed by default can be computed
in the same time as the count that --count used to produce.

If --quiet is passed, imgcmp exits with 1 if there's a difference,
but doesn't produce any output.

As before, if there are no differences, there's no output.