LibGfx/Tests: Make JBIG2 composition tests for "or" more thorough
To test for 'or', we painted a region twice in the same location,
on a white background.

But if 'or' was accidentally implemented with 'replace' semantics,
the test would still pass.

So draw a third, completely white region on top, also with the 'or'
operator.

Do this for the external combination operator tests, but also for the
internal combination_operator test for text regions.
(Text regions can't use "replace" as internal operator, but it still
seems good to test this there too.)

In the halftone composition test, fix a too-big region while here.
This was harmless, but it's tidier now.